Overhaul the init process
There are several new assumptions we are making that can drastically
simplify the init process, as well as a new optimization.

Assumptions:
- The script is only ever run at the very start of the game
- The game is paused and nothing changes other than what the script is
doing
- Towns are not being created or destroyed
- Industries are not being created or destroyed, except by the script

The init method now builds up in advance a list of all eligible towns
and tiles for industry construction. As the script proceeds, this list
of tiles will be winnowed down.

For reference, the old method would try towns at random, over and over,
even if a town has already failed to work.