Add in terrain filtering to town tile search
There's one big issue this raises. We're tracking town eligibility in a
single GSList, but different industries are and aren't buildable on
different terrains. For example, we might exhaust all the land tiles
around a town for building on -- but there might be plenty of water
tiles still available. In this case, if we were trying to build a
land-based industry, we would certainly say the town is no longer
eligible; but if we were building a water based industry, the town
is. Currently, our design does not respect this distinction.

I think the right solution is to put the terrain class in the TownBuild
function, and instead of tracking eligibility in a single list we'll
have to just track it in parallel based on terrain class. Yet another
helper function will be needed.