Correctly check industry terrain eligibility
This is a major rework of what we were doing before. Essentially,
whether or not a tile is eligible for an industry depends on both the
terrain of the tile and the industry itself. Some industries can only be
built on certain types of terrain.

So, instead of there being a single list of tiles with whether or not
they are eligible for construction, we have to maintain a list for each
terrain type. These lists may overlap for terrain types that overlap --
nonsnow nondesert, and nonsnow, for example. When we attempt to place a
random industry, we select the industry, then look up the terrain types
it wants, then query the relevant list for a tile to attempt to build
on. This tile is then removed from all the lists it may be in.

This will probably be extended again in the future, since these lists
are essentially 'mapmodes' for eligibility in the sense that they track
whether or not we are interested in trying to build an industry on those
tiles. As we extend the script, we can store other tile metadata in
lists like the height of each tile, whether that tile is a slope, etc.