Add in workaround for industry rate limiting
The way this works is by using the player company to build industries.

The other workaround here is for shore-based industries. Industries are
built from their top-left corner, so without the extra buffer of tiles
the only shore based industries that can and will be built will be shore
based industries where the shoreline is on the left or right of the industry.